+++ activerecord/lib/active_record/connection_adapters/postgresql/oid/date.rb
@@ -1,0 +1,11 @@
+module ActiveRecord
+  module ConnectionAdapters
+    module PostgreSQL
+      module OID # :nodoc:
+        class Date < Type::Date
+          include Infinity
+        end
+      end
+    end
+  end
+end
