+++ activerecord/lib/active_record/connection_adapters/postgresql/oid/date_time.rb
@@ -1,0 +1,26 @@
+module ActiveRecord
+  module ConnectionAdapters
+    module PostgreSQL
+      module OID # :nodoc:
+        class DateTime < Type::DateTime
+          include Infinity
+
+          def cast_value(value)
+            if value.is_a?(::String)
+              case value
+              when 'infinity' then ::Float::INFINITY
+              when '-infinity' then -::Float::INFINITY
+              when / BC$/
+                super("-" + value.sub(/ BC$/, ""))
+              else
+                super
+              end
+            else
+              value
+            end
+          end
+        end
+      end
+    end
+  end
+end
