+++ activerecord/lib/active_record/connection_adapters/postgresql/oid/inet.rb
@@ -1,0 +1,13 @@
+module ActiveRecord
+  module ConnectionAdapters
+    module PostgreSQL
+      module OID # :nodoc:
+        class Inet < Cidr
+          def type
+            :inet
+          end
+        end
+      end
+    end
+  end
+end
