+++ activerecord/lib/active_record/connection_adapters/postgresql/oid/infinity.rb
@@ -1,0 +1,13 @@
+module ActiveRecord
+  module ConnectionAdapters
+    module PostgreSQL
+      module OID # :nodoc:
+        module Infinity
+          def infinity(options = {})
+            options[:negative] ? -::Float::INFINITY : ::Float::INFINITY
+          end
+        end
+      end
+    end
+  end
+end
