+++ activerecord/lib/active_record/connection_adapters/postgresql/oid/specialized_string.rb
@@ -1,0 +1,19 @@
+module ActiveRecord
+  module ConnectionAdapters
+    module PostgreSQL
+      module OID # :nodoc:
+        class SpecializedString < Type::String
+          attr_reader :type
+
+          def initialize(type)
+            @type = type
+          end
+
+          def text?
+            false
+          end
+        end
+      end
+    end
+  end
+end
