+++ activerecord/lib/active_record/connection_adapters/postgresql/oid/time.rb
@@ -1,0 +1,11 @@
+module ActiveRecord
+  module ConnectionAdapters
+    module PostgreSQL
+      module OID # :nodoc:
+        class Time < Type::Time
+          include Infinity
+        end
+      end
+    end
+  end
+end
