+++ activerecord/lib/active_record/connection_adapters/postgresql/oid/uuid.rb
@@ -1,0 +1,17 @@
+module ActiveRecord
+  module ConnectionAdapters
+    module PostgreSQL
+      module OID # :nodoc:
+        class Uuid < Type::Value
+          def type
+            :uuid
+          end
+
+          def type_cast(value)
+            value.presence
+          end
+        end
+      end
+    end
+  end
+end
